window: fix notification not presenting the correct message
I've noticed that on the devel version there is a bug in the
notifications because when the pomodoro ends, it shows the "Back to
work" notification, and when the break ends, it shows the "Break time"
notification.

This commit fixes this after noticing that the wrong variable is being
passed to the `send_notifcation` function, because by looking at the
match done in that function, clearly it is expected to receive the kind
of lap that is about to start.

I've tested the before and after of this change to make sure that it is
working as expected.