window: Redesign using AdwToolbarView
libadwaita 1.4 provides the AdwToolbarView widget,
which provides a framework for toolbars and content.
AdwToolbarView defaults to a flat style, meaning we no
longer need to manually set `.flat` on the headerbar.
We also drop the GtkWindowHandle, as applications are
no longer expected to be draggable from everywhere.

Also includes slight padding changes and a new default
size.